alpha/requirements.txt
Flask~=3.0.3
Flask-SQLAlchemy~=3.1.1
Flask-Login~=0.6.3
Flask-WTF~=1.2.1
WTForms~=3.1.2
pillow~=10.4.0